requirements.txt
# PyTorch is installed separately (see scripts/setup.bat) with the CUDA 11.8 wheel.
# faiss-cpu replaces chromadb/chroma-hnswlib — pre-built wheels, no compiler needed.
faiss-cpu>=1.7.4
# sentence-transformers 3.0+ pulls transformers >= 4.40, sufficient for Qwen3.
sentence-transformers>=3.0.1
# Pinned floor for Qwen3-Embedding / Qwen3-Reranker config support.
transformers>=4.45
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
watchdog==4.0.1
python-dotenv==1.0.1
pydantic==2.7.1
tiktoken==0.7.0
rich==13.7.1
rank-bm25==0.2.2
# Multi-language AST chunking (TS/JS/Go/Rust/Java/C/C++/C#/PHP/Ruby/Swift/Kotlin).
# Bundled grammars download lazily on first parse.
tree-sitter-language-pack>=1.6.0
numpy>=1.24.0,<2.0
voyageai>=0.2.3
pytest==8.2.2
pytest-asyncio==0.23.7

# ── Optional: multi-graph (Kuzu embedded property graph) ─────────────────────
# Falls back silently to SQLite when not installed. Adds ~30 MB.
kuzu>=0.4.0; python_version >= "3.10"

# ── Optional: late-interaction retrieval (ColBERT/PyLate) ────────────────────
# Token-level retrieval. Disabled cleanly when not installed.
# pylate>=1.0.0   # uncomment to enable

# ── Optional: eval suite YAML loader ─────────────────────────────────────────
PyYAML>=6.0

# ── Optional: sandbox verifier toolchain (install per-language as needed) ────
# Python   : pip install ruff mypy
# TS/JS    : npm i -g typescript eslint
# Go       : gofmt + go vet ship with the Go toolchain
# Rust     : rustup default stable
# Java/C++ : javac, clang/gcc from your system toolchain
